Managing favorite Meals with the Context API
